feat: настроить страницу с сохранёнными фильмами
