+++ src/components/SavedMovies/SavedMovies.js
@@ -1,3 +1,13 @@
-export default function SavedMovies() {
-  return <article className="saved-movies-page"></article>;
+import MoviesCardList from "../MoviesCardList/MoviesCardList";
+import Preloader from "../Preloader/Preloader";
+import SearchForm from "../SearchForm/SearchForm";
+
+export default function SavedMovies(props) {
+  return (
+    <article className="saved-movies-page">
+      <SearchForm />
+      <MoviesCardList movies={props.movies} />
+      <Preloader />
+    </article>
+  );
 }
